% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sales_ts.R
\name{clean_sales_ts}
\alias{clean_sales_ts}
\title{Czyszczenie szeregów czasowych sprzedaży}
\usage{
clean_sales_ts(
  df,
  fill_na_method = "zero",
  time_aggregation = NULL,
  sort_data = TRUE
)
}
\arguments{
\item{df}{Obiekt tibble (wynik działania funkcji load_sales_data).}

\item{fill_na_method}{Metoda uzupełniania braków w sprzedaży: "zero" (domyślnie), "mean" lub "locf" (last observation carried forward).}

\item{time_aggregation}{Opcjonalna agregacja dat: NULL (brak, dane dzienne), "week", "month".}

\item{sort_data}{Wartość logiczna (TRUE/FALSE) - czy posortować dane chronologicznie.}
}
\value{
Oczyszczony obiekt tibble.
}
\description{
Funkcja przetwarza dane sprzedażowe: usuwa braki danych, obsługuje
duplikaty, umożliwia sortowanie oraz opcjonalną agregację czasową.
}
