% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_management_summary.R
\name{create_management_summary}
\alias{create_management_summary}
\title{Tworzenie podsumowania menedżerskiego}
\usage{
create_management_summary(df, period_days = 30)
}
\arguments{
\item{df}{Obiekt tibble (oczyszczone dane sprzedażowe).}

\item{period_days}{Liczba dni definiująca "ostatni okres" do analizy dynamiki (domyślnie 30 dni).}
}
\value{
Lista zawierająca zestaw kluczowych wskaźników efektywności (KPI) i wniosków biznesowych.
}
\description{
Funkcja generuje raport biznesowy: identyfikuje najlepszy i najgorszy sklep pod względem sprzedaży,
najszybciej rosnącą kategorię produktów, produkt o największym spadku procentowym w ostatnim okresie
oraz średnią sprzedaż w ostatnim okresie w porównaniu do historycznej.
}
