% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prognosis.R
\name{create_prognosis}
\alias{create_prognosis}
\title{Prognozowanie sprzedaży przy użyciu modeli ARIMA i Prophet}
\usage{
create_prognosis(df, h = 30)
}
\arguments{
\item{df}{Obiekt tibble (oczyszczone dane sprzedażowe).}

\item{h}{Liczba dni do przodu, dla których ma zostać wykonana prognoza (domyślnie 30).}
}
\value{
Lista zawierająca: 'arima_forecast' (obiekt prognozy ARIMA), 'prophet_forecast' (ramka danych z wynikami Prophet) oraz 'comparison_df' (tabela porównawcza prognoz).
}
\description{
Funkcja agreguje dane do całkowitej sprzedaży dziennej, a następnie tworzy prognozę
na określoną liczbę dni w przód za pomocą automatycznego modelu ARIMA oraz modelu Prophet.
}
