% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sales_data.R
\name{load_sales_data}
\alias{load_sales_data}
\title{Wczytywanie danych sprzedażowych}
\usage{
load_sales_data(path_to_dir)
}
\arguments{
\item{path_to_dir}{Ścieżka do folderu, w którym znajdują się pobrane pliki CSV.}
}
\value{
Obiekt tibble zawierający połączone dane z train.csv, stores.csv i holidays_events.csv.
}
\description{
Funkcja wczytuje surowe pliki CSV pobrane ze strony Kaggle
łączy je w jeden obiekt typu tibble, korzystając z pakietu tidyverse.
}
