% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sales_trends.R
\name{plot_sales_trends}
\alias{plot_sales_trends}
\title{Wizualizacja trendów sprzedaży}
\usage{
plot_sales_trends(df, group_var = NULL)
}
\arguments{
\item{df}{Obiekt tibble (oczyszczone dane sprzedażowe).}

\item{group_var}{Opcjonalna nazwa zmiennej jako ciąg znaków (np. "family" lub "store_nbr"), według której dane mają zostać pogrupowane na wykresie.}
}
\value{
Obiekt ggplot przedstawiający trendy sprzedaży.
}
\description{
Funkcja tworzy wykres liniowy przedstawiający trendy sprzedaży w czasie przy użyciu pakietu ggplot2.
Pozwala na opcjonalne grupowanie danych (np. po kategoriach produktów lub sklepach).
}
