% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sales_ts.R
\name{validate_sales_ts}
\alias{validate_sales_ts}
\title{Walidacja jakości szeregów czasowych}
\usage{
validate_sales_ts(df)
}
\arguments{
\item{df}{Obiekt tibble (najlepiej wynik działania funkcji load_sales_data).}
}
\value{
Lista zawierająca szczegółowy raport z weryfikacji danych.
}
\description{
Funkcja sprawdza jakość przygotowanych danych sprzedażowych. Weryfikuje braki,
duplikaty, poprawne daty, zakresy wartości oraz niespójną częstotliwość.
}
